Switch from calculating dihedrals to translations
In this new mode, the tranlation from each atom to the root atom of the
template is calculated instead of the dihedral angle
