+++ peleffy/BCEConformations/__init__.py
@@ -1,0 +1,1 @@
+from .BCEConformations import BCEConformations
