--- peleffy/BCEDihedrals/__init__.py
@@ -1,1 +1,0 @@
-from .BCEDihedrals import BCEDihedrals
